app2-oidc/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
authlib==1.3.0
requests==2.31.0
flask-session==0.5.0
redis==5.0.1
gunicorn==21.2.0
PyJWT==2.8.0
cryptography==41.0.7